extract the git back up code into its own file
The goal here is to be able to explore replaceing GitLibrary with SwiftGitX
